Propagate errors from get_available_kb()
It took me some time to identify a proper syntax to propagate the
error from a match branch, now we can be more specific about what we
ignore.  Now log a warning once in the "Unsupported" case (when
fetching the total memory), and let other memory-collector errors
cause a fatal error.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>