netlink: use name from NLA in Link events
We were always calling `if_indextoname()` to get the interface name when
receiving a New/DelLink message.  Just use the data we already received
instead.

On NewLink messages we just received the interface name, so we don't need
to make another query to get the information; on DelLink the sycall cannot
even give us this information any more.

Note at this point we still use "" to denote the lack of a name - it
seems unlikely to happen there anyway, and will be made cleaner next.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>