pnet: index NetworkView by interface index, not name
Current structure made us fill NetInterface with a dummy interface
number, and we want to use interface number in xenstore_schema_std to
fix #13.

Iface name becomes part of the associated data, bundle it into
InterfaceInfo for now (it will get redundant with introduction of a more
global interface cache).

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>