xenstore_schema_std: index the iface address cache with OS index
Iface name is not always part of a Netlink message, as opposed to the
OS index of the interface, and looking up the name following reception of
an interface removal notification just does not work.

All supported and prospective platforms seem to have a concept of
interface OS index, so this should cause no problem and be more efficient.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>