net: separate add/rm of interface and add/rm of MAC address
The current code would do nothing on interface removal, as we only
emit a `RmMac` event, with which xenstore_schema_std does nothing (we did
not add them in the first place).

An explicit `RmIface` allows for the removal of the full iface tree in
Xenstore, since there is no need to expose a mostly-empty tree.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>