net: keep a cache of network interfaces
Collects in a central place a map of iface indices, their names, and their
toolstack interface if any.

Finally replaces add_vif_info() calls with get_toolstack_interface()
called once to populate the cache.

Finally the RmIp events can be applied when an interface disappears, which
essentially fixes #13.

Though this makes a new problem apparent: recreating the VIF sometimes
does not get caught on Linux, when the iface is created as eth0 and
then renamed by udev as enX0 in a subsequent Newlink message: since we
are now only attempting to identify the toolkit-interface once, if
when we try to access the /sys node for eth0 the renaming already
occured the toolkit-interface is frozen as None.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>